rename grm to gcd | add all resources to the same objectid
