+++ maestro/main.go
@@ -37,7 +37,7 @@ import (
 	"github.com/armPelionEdge/maestro/maestroutils"
 	"github.com/armPelionEdge/maestro/mdns"
 	"github.com/armPelionEdge/maestro/networking"
-	"github.com/armPelionEdge/maestro/grm"
+	"github.com/armPelionEdge/maestro/gcd"
 	"github.com/armPelionEdge/maestro/processes"
 	"github.com/armPelionEdge/maestro/storage"
 	"github.com/armPelionEdge/maestro/sysstats"
@@ -451,7 +451,7 @@ func main() {
 		go bringUpIfs()
 	}
     // Gateway Resource Manager
-	go grm.Grm_init(config.Grm)
+	go gcd.Gcd_init(config.GatewayCapabilities)
 
 	/*********************************************/
 	/*               Jobs startup                */
