+++ maestroConfig/config.go
@@ -76,7 +76,7 @@ type YAMLMaestroConfig struct {
 	PlatformReaders      []PlatformReader                        `yaml:"platform_readers"`
 	Plugins              []Plugin                                `yaml:"plugins"`
 	Network              *maestroSpecs.NetworkConfigPayload      `yaml:"network"`
-	Grm                  *GrmConfig                              `yaml:"grm`
+	GatewayCapabilities  *GcdConfig                              `yaml:"gateway_capabilities`
 	Processes            *configs.ProcessesConfig                `yaml:"processes"`
 	DebugOpts            *DebugOptions                           `yaml:"debug_opts"`
 	DDBConnConfig        *DeviceDBConnConfig                     `yaml:"devicedb_conn_config"`
@@ -97,12 +97,18 @@ type DebugOptions struct {
 	PidFileDates bool `yaml:"pid_file_dates"`
 }
 
-type GrmConfig struct {
+type GcdConfig struct {
 	EdgeCoreSocketPath string `yaml:"edge_core_socketpath"`
-	FluentbitConfigFilePath string `yaml:"fluentbit_config_filepath"`
-	FluentbitConfigObjectId int `yaml:"fluentbit_config_lwm2mobjectid"`
-	
+	ConfigObjectId int `yaml:"lwm2m_objectid"`
+	GatewayResources []GatewayResource `yaml:"gateway_resources"`
 }
+
+type GatewayResource struct {
+	Name string `yaml:"name"`
+	Enable bool `yaml:"enable"`
+	ConfigFilePath string `yaml:"config_filepath"`
+}
+
 type DeviceDBConnConfig struct {
 	// The URI of the relay's local DeviceDB instance
 	DeviceDBUri string `yaml:"devicedb_uri"`
